+++ shopping_list/src/App.vue
@@ -3,6 +3,7 @@
 		<NcAppNavigation>
 			<ListSidebar @show-settings="showSettings = true" />
 			<template #footer>
+				<AndroidAppLink />
 				<ItemImagesSettings />
 			</template>
 		</NcAppNavigation>
@@ -35,6 +36,7 @@ import ListSidebar from './components/ListSidebar.vue'
 import ListView from './components/ListView.vue'
 import AreaKeywordsSettings from './components/AreaKeywordsSettings.vue'
 import ItemImagesSettings from './components/ItemImagesSettings.vue'
+import AndroidAppLink from './components/AndroidAppLink.vue'
 import OfflineIndicator from './components/OfflineIndicator.vue'
 import { useListsStore } from './stores/lists'
 import { usePush } from './composables/usePush'